requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
distro==1.9.0
Django==5.1.3
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-filter==24.3
django-modeltranslation==0.19.11
django-rosetta==0.10.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.11.1
drf-yasg==1.21.8
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
inflection==0.5.1
jiter==0.7.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
packaging==24.2
pillow==11.0.0
polib==1.2.0
psycopg2-binary==2.9.10
PyJWT==2.10.0
PyPDF2==3.0.1
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.21.0
sniffio==1.3.1
sqlparse==0.5.2
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
